+++ FrontEndApp/v1-react/src/resources/keras/kLayers.js
